Extract infinite scroll logic in own component
